Authorization and dockblocks

+++ app/Http/Controllers/BookController.php
@@ -9,18 +9,31 @@ use Symfony\Component\HttpFoundation\Response;
 
 class BookController extends Controller
 {
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
     public function index()
     {
-        // $this->authorize('viewAny', Book::class);
+        $this->authorize('viewAny', Book::class);
 
         $books = Book::with('author')->paginate();
 
-        return BookResource::collection($books);
+        return (BookResource::collection($books))
+            ->response()
+            ->setStatusCode(Response::HTTP_PARTIAL_CONTENT);
     }
 
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  App\Http\Requests\StoreBookRequest $request
+     * @return \Illuminate\Http\Response
+     */
     public function store(StoreBookRequest $request)
     {
-        // $this->authorize('create');
+        $this->authorize('create', Book::class);
 
         $book = Book::create($request->validated());
 
@@ -29,16 +42,29 @@ class BookController extends Controller
             ->setStatusCode(Response::HTTP_CREATED);
     }
 
+    /**
+     * Display the specified resource.
+     *
+     * @param  App\Book  $book
+     * @return \Illuminate\Http\Response
+     */
     public function show(Book $book)
     {
-        // $this->authorize('view', $book);
+        $this->authorize('view', $book);
 
         return BookResource::make($book->load('author'));
     }
 
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  App\Book  $book
+     * @param  App\Http\Requests\StoreBookRequest $request
+     * @return \Illuminate\Http\Response
+     */
     public function update(Book $book, StoreBookRequest $request)
     {
-        // $this->authorize('update', $bookRecord);
+        $this->authorize('update', $book);
 
         $book->update($request->validated());
 
@@ -47,9 +73,15 @@ class BookController extends Controller
             ->setStatusCode(Response::HTTP_OK);
     }
 
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  App\Book  $book
+     * @return \Illuminate\Http\Response
+     */
     public function destroy(Book $book)
     {
-        // $this->authorize('delete', $book);
+        $this->authorize('delete', $book);
 
         $book->delete();
